fix(api): address team user review feedback
